% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_as_xlsx.R
\name{ae_as_xlsx}
\alias{ae_as_xlsx}
\title{Export AE table to .xlsx}
\usage{
ae_as_xlsx(ae_tbl, file, bold_rows = NULL, header_rows = NULL)
}
\arguments{
\item{ae_tbl}{A 'tbl_ae_count' object}

\item{file}{Export destination and name for the .xlsx file}

\item{bold_rows}{An optional numeric vector of rows to bold}

\item{header_rows}{An optional numeric vector of header row indices}
}
\value{
Exports a .xlsx file
}
\description{
Quickly export a 'tbl_ae_count' object to an Excel file with
prespecified formatting that approximates the look of the corresponding
.html.
}
\examples{
df_adverse_events \%>\%
tbl_ae(
  id = patient_id,
  ae = adverse_event,
  soc = system_organ_class,
  by = grade
) \%>\%
  ae_as_xlsx("ae-table.xlsx", bold_rows = 1:2, header_rows = 1:2)
}
