% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_table_shell.R
\name{as_table_shell}
\alias{as_table_shell}
\title{Create Table Shell}
\usage{
as_table_shell(
  x,
  columns = c(gtsummary::all_stat_cols(), dplyr::any_of(c("p.value", "q.value",
    "estimate", "ci", "conf.low", "conf.high", "statistic"))),
  pattern = "[[:digit:]]+",
  replacement = "xx",
  replace_headers = NULL,
  replace_spanning_headers = TRUE
)
}
\arguments{
\item{x}{a 'gtsummary' object}

\item{columns}{columns to replace digits with placeholder}

\item{pattern}{regex pattern of text to replace. Default is \code{"[[:digit:]]+"}}

\item{replacement}{string placeholder. Default is \code{"x"}}

\item{replace_headers}{logical indicating whether to replace digits in column headers.
Default is FALSE for AE tables and TRUE for all other tables.}

\item{replace_spanning_headers}{logical indicating whether to replace digits
in column spanning headers. Default is TRUE}
}
\value{
a 'gtsummary' table
}
\description{
Convert a 'gtsummary' table into a table shell by replacing the statistics
reported in the table with placeholders.
}
\examples{
tbl <-
  df_adverse_events \%>\%
  tbl_ae(
    id = patient_id,
    ae = adverse_event,
    soc = system_organ_class,
    by = grade,
    zero_symbol = NULL
  ) \%>\%
  modify_header(all_ae_cols() ~ "**Grade {by}**") \%>\%
  as_table_shell()
}
