% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_check.R
\name{beta_check}
\alias{beta_check}
\title{Beta Check}
\usage{
beta_check(bmat, Xs, nXs, nYS)
}
\arguments{
\item{nYS}{}
}
\value{
A list containing the minimum of Beta2(X) and the vector of Beta2(X).
}
\description{
This function computes the values for Beta2(X) and returns them highlighting the minimum value.
}
