% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_coef.R
\name{calibrate_coef}
\alias{calibrate_coef}
\title{Calibration Function}
\usage{
calibrate_coef(Y, X, qY = NULL, dgp = "dr", link = "probit")
}
\arguments{
\item{link}{}
}
\value{

}
\description{
This function calibrates coefficients.
}
