% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulfunc.R
\name{califunc}
\alias{califunc}
\title{califunc}
\usage{
califunc(Y, X, qY = NULL, dgp = "dr", link = "probit")
}
\arguments{
\item{link}{}
}
\value{

}
\description{
This function is for simulations. It calibrates the coefficients.
}
