% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_contract.R
\name{cm_contract}
\alias{cm_contract}
\title{CM Contract Function}
\usage{
cm_contract(cm, buse)
}
\arguments{
\item{buse}{}
}
\value{

}
\description{
This contracts a 'contributions' (n by big) matrix by the mask in buse, by removing the appropriate columns.
}
