% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcontract.R
\name{cmcontract}
\alias{cmcontract}
\title{CM Contract Function}
\usage{
cmcontract(cm, buse)
}
\arguments{
\item{buse}{}
}
\value{

}
\description{
This contracts a 'contributions' (n by bbig) matrix by the mask in buse, by removing the appropriate columns.
}
