% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_data.R
\name{complete_data}
\alias{complete_data}
\title{Create a complete and expanded data frame for tabulating adverse events}
\usage{
complete_data(data, id, ae, soc, by, strata, id_values, by_values)
}
\arguments{
\item{data}{Data frame}

\item{id}{Variable name of the patient ID}

\item{ae}{Variable name of the adverse event column}

\item{soc}{Variable name of the system organ class column}

\item{by}{Variable to split results by, e.g. report AEs by grade or attribution}

\item{strata}{Variable to stratify results by, e.g. report AEs summaries
by treatment group}

\item{id_values}{Optional vector of complete id values to achieve correct
base n for the situation in which not all subjects experience adverse events}

\item{by_values}{Optional vector of complete by values, listed in desired order,
to achieve correct table structure for the situation in which an adverse
event of a certain grade is not observed for a given soc}
}
\description{
Create a complete and expanded data frame for tabulating adverse events
}
\examples{
df_aes \%>\%
  tbl_adverse_events(
    id = patient_id,
    ae = adverse_event,
    soc = system_organ_class,
    by = grade,
    strata = trt
  ) \%>\%
  as_kable() # UPDATE THIS WITH PROPER gt image at some point..
}
