% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep4.R
\name{cqf.func4}
\alias{cqf.func4}
\title{cqf.func4}
\usage{
cqf.func4(
  y,
  x,
  xgrid,
  info,
  Ysing = FALSE,
  bmat,
  ng.plot,
  ydf,
  yorder,
  ugrid,
  nr,
  nc,
  delta = 1
)
}
\arguments{
\item{delta}{}
}
\value{

}
\description{
This function is internal.
}
