% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep4.R
\name{data.info}
\alias{data.info}
\title{data.info}
\usage{
data.info(
  x,
  X.type,
  xdf,
  xknots = NULL,
  coord.bare = NULL,
  coord.spline = NULL,
  coord.tensor = NULL,
  nxgrid = 101,
  gridx.cont = F
)
}
\arguments{
\item{gridx.cont}{}
}
\value{

}
\description{
This function is internal. It generates the data info for a variable.
}
