% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{data_info}
\alias{data_info}
\title{Data Info Function}
\usage{
data_info(
  x,
  X.type,
  xdf,
  xknots = NULL,
  coord.bare = NULL,
  coord.spline = NULL,
  coord.tensor = NULL,
  nxgrid = 101,
  gridx.cont = F,
  delta.ok = F
)
}
\arguments{
\item{delta.ok}{}
}
\value{
The information for this data.
}
\description{
This function creates the necessary info for the data.
}
