% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norm.R
\name{data_norm}
\alias{data_norm}
\title{Data Normalisation}
\usage{
data_norm(x, mean = 0, sd = 1, log = FALSE)
}
\arguments{
\item{x}{The data to be normalised.}

\item{mean}{The value to center on.}

\item{sd}{The standard deviation to target.}

\item{log}{Whether to log the data. NOT IMPLEMENTED}
}
\value{
It returns a normalised version of the data.
}
\description{
This function normalises the data.
}
