% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Data Preparation}
\usage{
data_prep(
  y,
  x,
  info = NULL,
  ygrid = NULL,
  xgrid = NULL,
  bmat = NULL,
  y_knots = NULL,
  addxint = T,
  ydf,
  yorder,
  yorth = FALSE,
  xorth = FALSE,
  Ysing = FALSE,
  e0mode = F,
  mask = NULL,
  returnTZ = T,
  returnTZg = F,
  de0dygrid = NULL,
  extrapolate = T,
  plot.mode = F,
  delta = 1,
  nxgrid,
  nygrid,
  Ys = NULL,
  YS = NULL,
  Xs = NULL,
  Xsgrid = NULL,
  sYgrid = NULL,
  SYgrid = NULL,
  TZ = NULL,
  tZ = NULL,
  TZ.pl = NULL,
  tZ.pl = NULL,
  TZgrid = NULL,
  tZgrid = NULL,
  TZgrid.pl = NULL,
  tZgrid.pl = NULL
)
}
\arguments{
\item{tZgrid.pl}{}
}
\value{
The design matrices for Y and X.
}
\description{
A function that takes Y and X and returns design matrices
}
