% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulate.R
\name{data_simulate}
\alias{data_simulate}
\title{Data Simulation Function.}
\usage{
data_simulate(
  Ydata,
  Xdata,
  bvec,
  n = 100,
  sde = 1,
  dgp = "dr",
  ugrid = NULL,
  xgrid = NULL
)
}
\arguments{
\item{xgrid}{}
}
\value{
Some fake data.
}
\description{
This function generates some fake data.
}
