% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_adverse_events}
\alias{df_adverse_events}
\title{Simulated Adverse Event Database}
\format{
A data frame with 100 rows--one row per patient per AE
\describe{
\item{patient_id}{Patient ID}
\item{trt}{Treatment Group}
\item{system_organ_class}{System Organ Class}
\item{adverse_event}{Adverse Event}
\item{grade}{Grade}
\item{drug_attribution}{Drug Attribution}
\item{any_complication}{Any Grade Complication}
\item{grade3_complication}{Grade 3+ Complication}
}
}
\usage{
df_adverse_events
}
\description{
A data set containing reported AEs from a trial.
}
\keyword{datasets}
