% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_patient_characteristics}
\alias{df_patient_characteristics}
\title{Simulated Patient Characteristics Database}
\format{
A data frame with 100 rows--one row per patient
\describe{
\item{patient_id}{Patient ID}
\item{trt}{Treatment Group}
\item{age}{Patient Age}
\item{marker}{Biological Marker}
\item{status}{Study Status}
\item{discontinued}{Discontinued from Study}
\item{off_trt_ae}{Off Treatment Adverse Event}
}
}
\usage{
df_patient_characteristics
}
\description{
Simulated Patient Characteristics Database
}
\keyword{datasets}
