% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linebreak_gtsummary.R
\name{.linebreak_gtsummary}
\alias{.linebreak_gtsummary}
\title{Make linebreak in LaTeX Table cells with gtsummary}
\usage{
.linebreak_gtsummary(
  x,
  align = NULL,
  double_escape = FALSE,
  linebreaker = "\\n"
)
}
\arguments{
\item{x}{a gtsummary table}

\item{align}{instruction passed to \code{kableExtra::linebreak(align=)}. Default
is NULL, which pulls alignment instructions from the gtsummary object.}

\item{double_escape}{Whether special character should be double escaped.
Default is FALSE.}

\item{linebreaker}{Symbol for linebreaks to replace. Default is \verb{\\\\n}.}
}
\value{
string of column names in LaTeX format
}
\description{
This function creates a call to \code{kableExtra::linebreak()}, and is meant to used in
\code{as_kable_extra(x, col.names = .linebreak_gtsummary(x), escape = FALSE, format = "latex")}.
Note that \code{escape = FALSE} and \code{format = "latex"} are required.
\itemize{
\item the default \verb{align=} argument is taken from the gtsummary object
\item the markdown double-star and double-underscore bold syntax is converted to LaTeX, \verb{\\textbf\{\}}
\item the markdown single-star and single-underscore italic syntax is converted to LaTeX, \verb{\\textit\{\}}
}
}
\examples{
# Example 1 -----------------------------------------------------------------
df_adverse_events \%>\%
  tbl_ae(
    id = patient_id,
    ae = adverse_event,
    by = grade,
    header_by = "**Grade {level}**"
  ) \%>\%
  as_kable_extra(
    col.names = .linebreak_gtsummary(.),
    escape = FALSE,
    format = "latex"
  )
}
