% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_helper.R
\name{eb_helper}
\alias{eb_helper}
\title{Evaluation of Basis Functions Helper}
\usage{
eb_helper(x, M, knots, order)
}
\arguments{
\item{order}{}
}
\value{

}
\description{
This function helps to evluate the basis functions.
}
