% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_basis.R
\name{eval_basis}
\alias{eval_basis}
\title{Evaluation of Basis Functions.}
\usage{
eval_basis(object, x, ...)
}
\arguments{
\item{...}{}
}
\description{
This function evluates the basis functions. It is based closely on evaluate method for spline basis in package orthogonalsplinebasis, but about 30% faster.
}
