% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_basis.R
\name{EvalBasis}
\alias{EvalBasis}
\title{EvalBasis Function}
\usage{
EvalBasis(object, x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
This function evluates the basis functions. It is based closely on evaluate method for spline basis in package orthogonalsplinebasis, but about 30% faster.
}
