% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique.R
\name{get_unique}
\alias{get_unique}
\title{Get unique values from a variable in a data frame}
\usage{
get_unique(data, var, drop_na = TRUE)
}
\arguments{
\item{data}{data frame}

\item{var}{variable from which we want unique values}

\item{drop_na}{whether or not to drop NA from returned values; default to TRUE}
}
\value{
a vector (either character or numeric, depending on input)
}
\description{
Values are returned in their respective natural ordering, whether it be numeric,
character, or factor. If the input variable is a factor, get_unique()
removes factor levels and returns a character vector.
}
\examples{
dat <- data.frame(
  x = c(1, 1, 2),
  y = c("B", "B", "A"),
  z = factor(x = c("cat", "cat", "dog"), levels = c("dog", "cat"))
  )
get_unique(dat, x)
get_unique(dat, y)
get_unique(dat, z)

}
