% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep4.R
\name{grids.func}
\alias{grids.func}
\title{grids.func}
\usage{
grids.func(x, X.type, nxgrid = 101, gridx.cont = F)
}
\arguments{
\item{gridx.cont}{}
}
\value{

}
\description{
This function is internal. It generates the grids for a variable.
}
