% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids_func.R
\name{grids_func}
\alias{grids_func}
\title{Grids Function}
\usage{
grids_func(x, X.type, nxgrid = 101, grid.user = NULL, gridx.cont = F)
}
\arguments{
\item{x}{The explanatory data.}

\item{gridx.cont}{}
}
\description{
This generates the grids of values for the explanatory variable.
}
