% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_normalise.R
\name{gtr_normalise}
\alias{gtr_normalise}
\title{GTR Normalise}
\usage{
gtr_normalise(
  y,
  nyg.max = 0,
  plotting = F,
  eps = 1e-05,
  ygrid = NULL,
  log.t = F,
  doprimal = F,
  gam0 = 0.1
)
}
\arguments{
\item{gam0}{}
}
\description{
This function performs a transformation to normality, or at least to the real line.
}
