% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_solve_py.R
\name{gtr_solve_py}
\alias{gtr_solve_py}
\title{GTR Solver Using python Optimisation}
\usage{
gtr_solve_py(TYX, tYX)
}
\arguments{
\item{tYX}{}
}
\value{
The dictionary with the optimisation results.
}
\description{
This function solves the optimisation problem using the python package.
}
