% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtreg-package.R
\docType{package}
\name{gtreg-package}
\alias{gtreg}
\alias{gtreg-package}
\title{gtreg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shannonpileggi.github.io/gtreg}
}

}
\author{
\strong{Maintainer}: Shannon Pileggi \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})
}

}
\keyword{internal}
