% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx6_bare_data_frame.R
\name{gX6.bare.data.frame}
\alias{gX6.bare.data.frame}
\title{gX6 Data Frame (Bare)}
\usage{
\method{gX6}{bare.data.frame}(info, X)
}
\arguments{
\item{X}{}
}
\description{
This is the bare dataframe function version for gX6.
}
