% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx6_bare_matrix.R
\name{gX6.bare.matrix}
\alias{gX6.bare.matrix}
\title{gX6 Matrix (Bare)}
\usage{
\method{gX6}{bare.matrix}(info, X)
}
\arguments{
\item{X}{}
}
\value{

}
\description{
This is the bare matrix function version for gX6.
}
