% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx6_data_frame.R
\name{gX6.data.frame}
\alias{gX6.data.frame}
\title{gX6 Data Frame (Main)}
\usage{
\method{gX6}{data.frame}(X, info, orth = FALSE)
}
\arguments{
\item{orth}{}
}
\description{
This is the main dataframe function for gX6.
}
