% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx6_matrix.R
\name{gX6.matrix}
\alias{gX6.matrix}
\title{gX6 Matrix (Main)}
\usage{
\method{gX6}{matrix}(X, info, orth = FALSE)
}
\arguments{
\item{orth}{}
}
\value{

}
\description{
This is the main matrix function version for gX6.
}
