% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx6_spline_matrix.R
\name{gX6.spline.matrix}
\alias{gX6.spline.matrix}
\title{gX6 Matrix (Spline)}
\usage{
\method{gX6}{spline.matrix}(info, X, orth = FALSE, xord = 4)
}
\arguments{
\item{xord}{}
}
\value{

}
\description{
Transforms the data in X according to the instructions in info. Sees if there is a spline; if not just apply shift, scale.
}
