% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx6_spline_data_frame.R
\name{gX6_spline.data.frame}
\alias{gX6_spline.data.frame}
\title{gX6 Data Frame (Spline)}
\usage{
gX6_spline.data.frame(info, X, orth, xord = 4)
}
\arguments{
\item{xord}{}
}
\value{

}
\description{
This is the spline dataframe function version for gX6.
}
