% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Simple Hello Function}
\usage{
hello(name)
}
\arguments{
\item{name}{This is just someone's name.}
}
\value{
It returns a simple welcome message.
}
\description{
This function is just a test function to have a template of what the file for a function should look like.
}
\examples{
hello("Alex")
}
