% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text_tbl_ae.R
\name{inline_text_tbl_ae}
\alias{inline_text_tbl_ae}
\alias{inline_text.tbl_ae}
\alias{inline_text.tbl_ae_count}
\alias{inline_text.tbl_ae_focus}
\title{Report Values from \code{tbl_ae()} in-line}
\usage{
\method{inline_text}{tbl_ae}(x, ae_or_soc, column, ...)

\method{inline_text}{tbl_ae_count}(x, ae_or_soc, column, ...)

\method{inline_text}{tbl_ae_focus}(x, ae_or_soc, column, ...)
}
\arguments{
\item{x}{an object of class \code{tbl_ae()}, \code{tbl_ae_count()}, \code{tbl_ae_focus()}}

\item{ae_or_soc}{string indicating the AE or SOC to report}

\item{column}{column name of cell to report. Use \code{show_header_names(x)}
to print all column names beside the current header.}

\item{...}{not used}
}
\value{
string
}
\description{
Report Values from \code{tbl_ae()} in-line
}
\examples{
# add example
}
