% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_gtr.R
\name{inner.gtr.c}
\alias{inner.gtr.c}
\title{Inner GTR Computation Function}
\usage{
inner.gtr.c(
  TYX,
  tYX,
  y,
  x,
  nyg = 0,
  ng.qgm,
  weights = 1,
  zeros = NULL,
  reltol = 0.001,
  feastol = 0.001,
  abstol = 0.001,
  pen = NULL,
  cvg.mono = NULL,
  res.sol = NULL,
  fac = 1,
  fac.now2 = NULL,
  Xs = NULL,
  sYgrid = NULL,
  xgrid.qgm = NULL,
  Xs.qgm = NULL,
  sYgrid.qgm = NULL,
  info,
  yorder,
  ydf,
  iyknots = NULL,
  Ysing,
  maxit,
  nXs,
  nYS,
  lam.vec = NULL,
  gam = 0,
  ugrid = NULL,
  dedy.min = NULL,
  doprimal = F,
  tol.res,
  bounded = F,
  Cbound = Inf,
  cval = 0.1,
  algor = "ECOS",
  easy = T,
  threshold = 1e-05,
  e0mode = F
)
}
\arguments{
\item{e0mode}{}
}
\value{

}
\description{
This function does the inner computation for GTR.
}
