% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iyknots_func.R
\name{iyknots.func}
\alias{iyknots.func}
\title{IY Knots Function}
\usage{
iyknots.func(y, e0mode = F, ydf, yorder, delta = 1)
}
\arguments{
\item{delta}{}
}
\value{

}
\description{
This function generates the knots for the dependent variable.
}
