% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list.R, R/model_list2.R
\name{model_list}
\alias{model_list}
\title{Model List Function}
\usage{
model_list(yorder.dex, ydf.dex.0, xdf.dex)

model_list(yorder.dex, ydf.dex.0, xdf.dex)
}
\arguments{
\item{xdf.dex}{}
}
\value{



}
\description{
Model List Function

Model list2 function
}
