% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d_gtr.R
\name{plot3d.gtr}
\alias{plot3d.gtr}
\title{Plot3d GTR}
\usage{
plot3d.gtr(
  y,
  x,
  ugrid = seq(0.1, 0.9, by = 0.1),
  res,
  bmat,
  type = "CDF",
  ydf = NULL,
  yorder = NULL,
  info = NULL,
  Ysing = FALSE,
  ng.plot = 30,
  ygrid = NULL,
  xgrid = NULL,
  dx = 0,
  dy = 0,
  FOV = 1,
  alpha = 1,
  ncolors = 500,
  coltype = 1,
  xlab = "X",
  ylab = "Y",
  zlab = "",
  theta = 70,
  phi = 30,
  d = 1,
  expand = 1,
  r = sqrt(3),
  liness = F,
  shade = NA,
  blackgrid = NULL,
  projrq = F,
  BW = NULL,
  cex = 1,
  ticktype = "simple",
  sizept = 3,
  colpt = "blue",
  colines = "red",
  lwd.liness = 1,
  lwd.projrq = 1,
  col.liness = 1,
  col.projrq = 1,
  col.surface = "lightblue",
  zlim = NULL,
  main = NULL,
  delta = 1
)
}
\arguments{
\item{delta}{}
}
\value{
A plot of the gtr output with a scalar X.
}
\description{
This function is only for plotting simple gtr output with a scalar X for now.
}
