% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep4.R
\name{QGM.check}
\alias{QGM.check}
\title{QGM.check}
\usage{
QGM.check(
  y,
  x,
  res,
  xgrid.qgm = seq(min(x), max(x), len = 101),
  ugrid,
  nxgrid = 101,
  nygrid,
  ng.qgm = 201,
  nXs,
  nYS,
  ydf,
  yorder,
  iyknots = NULL,
  info = info,
  Ysing = FALSE,
  delta = 1,
  easy = F,
  zeros = NULL
)
}
\arguments{
\item{zeros}{}
}
\value{

}
\description{
This checks whether the solution that GTR provides satisfies Monotonicity. It creates a grid for the continous variables and checks for each value of discrete values.
}
