% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulfunc.R
\name{simdata}
\alias{simdata}
\title{simdata}
\usage{
simdata(
  Ydata,
  Xdata,
  bvec,
  n = 100,
  sde = 1,
  dgp = "dr",
  ugrid = NULL,
  xgrid = NULL
)
}
\arguments{
\item{xgrid}{}
}
\value{

}
\description{
This function is for simulations. It generates fake data that simulates the one provided.
}
