% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_xxx.R
\name{style_xxx}
\alias{style_xxx}
\title{Style numbers as x's}
\usage{
style_xxx(x, width = digits + 2, digits = 0)
}
\arguments{
\item{x}{a numeric or character vector}

\item{width}{the width of output field of x's, including the decimal place}

\item{digits}{the number of digits displayed after the decimal place}
}
\value{
a character vector
}
\description{
The purpose of \code{style_xxx()} is to convert numeric values in
summary tables to x's of consistent length for mock tables.
See the \href{https://shannonpileggi.github.io/gtreg/articles/table-shells.html}{Table shells vignette}
for detailed examples.
}
\examples{
style_xxx(7:10, digits = 0)
style_xxx(7:10, digits = 1)
style_xxx(7:10, width = 2, digits = 0)
style_xxx(7:10, width = 5, digits = 2)

}
