% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_adverse_event.R
\name{tbl_adverse_event}
\alias{tbl_adverse_event}
\title{Tabulate Adverse Events}
\usage{
tbl_adverse_event(
  data,
  id,
  ae,
  soc = NULL,
  by = NULL,
  strata = NULL,
  statistic = "{n} ({p})",
  header = "**{level}**"
)
}
\arguments{
\item{data}{Data frame}

\item{id}{Variable name of the patient ID}

\item{ae}{Variable name of the adverse event column}

\item{soc}{Variable name of the system organ class column}

\item{by}{Variable to split results by, e.g. report AEs by grade}

\item{strata}{Variable to stratify results by, e.g. report AEs summaries
by treatment group}

\item{statistic}{String indicating the statistics that will be reported.
The default is \code{"{n} ({p})"}}

\item{header}{String indicating the header to be placed in the table.
Default is \code{"**{level}**"}}
}
\description{
The function tabulates adverse events. One AE per ID will be counted in the
resulting table. If a \verb{by=} variable is passed and a
patient experienced more than one of the same AE, the AE associated with the
highest \verb{by=} level will be included. For example, if a patient has two of
the same AE and \code{by = grade}, the AE with the highest grade will be
included.
Similarly, if tabulations within system organ class are requested, the
AE within SOC associated with the highest grade will be tabulated.
}
\examples{
# Example 1 -----------------------------------------------------------------
df_adverse_events \%>\%
  tbl_adverse_event(
    id = patient_id,
    ae = adverse_event,
    soc = system_organ_class,
    by = grade,
    strata = trt,
    header = "**Grade {level}**"
  ) \%>\%
  as_kable() # UPDATE THIS WITH PROPER gt image at some point.

# Example 2 -----------------------------------------------------------------
df_adverse_events \%>\%
  tbl_adverse_event(
    id = patient_id,
    ae = adverse_event,
    by = grade,
    header = "**Grade {level}**"
  ) \%>\%
  as_kable() # UPDATE THIS WITH PROPER gt image at some point.
}
