% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_adverse_events.R
\name{tbl_adverse_events}
\alias{tbl_adverse_events}
\title{Tabulate Adverse Events}
\usage{
tbl_adverse_events(
  data,
  id,
  ae,
  soc = NULL,
  by = NULL,
  strata = NULL,
  statistic = "{n} ({p})",
  header = "**{level}**"
)
}
\arguments{
\item{data}{Data frame}

\item{id}{Variable name of the patient ID}

\item{ae}{Variable name of the adverse event column}

\item{soc}{Variable name of the system organ class column}

\item{by}{Variable to split results by, e.g. report AEs by grade}

\item{strata}{Variable to stratify results by, e.g. report AEs summaries
by treatment group}

\item{statistic}{String indicating the statistics that will be reported.
The default is \code{"{n} ({p})"}}

\item{header}{String indicating the header to be placed in the table.
Default is \code{"**{level}**"}}
}
\description{
Tabulate Adverse Events
}
\examples{
df_adverse_events \%>\%
  tbl_adverse_events(
    id = patient_id,
    adverse_event = adverse_event,
    soc = system_organ_class,
    grade = grade,
    strata = trt
  ) \%>\%
  as_kable() # UPDATE THIS WITH PROPER gt image at some point..
}
