% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ae.R
\name{tbl_ae}
\alias{tbl_ae}
\title{Tabulate Adverse Events}
\usage{
tbl_ae(
  data,
  id,
  ae,
  soc = NULL,
  by = NULL,
  strata = NULL,
  id_df = NULL,
  by_values = NULL,
  missing_text = "Unknown",
  statistic = "{n} ({p})",
  header = "**{level}**"
)
}
\arguments{
\item{data}{Data frame}

\item{id}{Variable name of the patient ID}

\item{ae}{Variable name of the adverse event column}

\item{soc}{Variable name of the system organ class column}

\item{by}{Variable to split results by, e.g. report AEs by grade}

\item{strata}{Variable to stratify results by, e.g. report AEs summaries
by treatment group}

\item{id_df}{Optional data frame of complete id values and strata to achieve correct
base n for the situation in which not all subjects experience adverse events}

\item{by_values}{Optional vector of complete by values, listed in desired order,
to achieve correct table structure for the situation in which an adverse
event of a certain grade is not observed for a given soc}

\item{missing_text}{String that will be shown for missing levels of \verb{by=},
Default is \code{"Unknown"}}

\item{statistic}{String indicating the statistics that will be reported.
The default is \code{"{n} ({p})"}}

\item{header}{String indicating the header to be placed in the table.
Default is \code{"**{level}**"}}
}
\description{
The function tabulates adverse events. One AE per ID will be counted in the
resulting table. If a \verb{by=} variable is passed and a
patient experienced more than one of the same AE, the AE associated with the
highest \verb{by=} level will be included. For example, if a patient has two of
the same AE and \code{by = grade}, the AE with the highest grade will be
included.
Similarly, if tabulations within system organ class are requested, the
AE within SOC associated with the highest grade will be tabulated.
}
\examples{
\dontshow{if (identical(Sys.getenv("NOT_CRAN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example 1 -----------------------------------------------------------------
df_adverse_events \%>\%
  tbl_ae(
    id = patient_id,
    ae = adverse_event,
    soc = system_organ_class,
    by = grade,
    strata = trt,
    header = "**Grade {level}**"
  ) \%>\%
  as_kable() # UPDATE THIS WITH PROPER gt image at some point.

# Example 2 -----------------------------------------------------------------
df_adverse_events \%>\%
  tbl_ae(
    id = patient_id,
    ae = adverse_event,
    by = grade,
    header = "**Grade {level}**"
  ) \%>\%
  as_kable() # UPDATE THIS WITH PROPER gt image at some point.
\dontshow{\}) # examplesIf}
}
