% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ae_binary.R
\name{tbl_ae_binary}
\alias{tbl_ae_binary}
\title{Tabluate AE Binary Summaries}
\usage{
tbl_ae_binary(
  data,
  include,
  id,
  ae,
  soc = NULL,
  strata = NULL,
  statistic = "{n} ({p})",
  include_label = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{include}{Vector of column names to summarize. Column names may be
quoted or unquoted. All columns must be class 'logical'.}

\item{id}{Variable name of the patient ID}

\item{ae}{Variable name of the adverse event column}

\item{soc}{Variable name of the system organ class column}

\item{strata}{Variable to stratify results by, e.g. report AEs summaries
by treatment group}

\item{statistic}{String indicating the statistics that will be reported.
The default is \code{"{n} ({p})"}}

\item{include_label}{A named list of labels that will be applied in the
resulting table. Names must be those passed in \verb{include=}. Default is
NULL, and either the label attribute or the column name will be used.}
}
\value{
a 'tbl_ae_binary' object
}
\description{
Tabluate AE Binary Summaries
}
\examples{
# Example 1 -----------------------------------------------------------------
tbl_ae_binary_ex1 <-
  df_adverse_events \%>\%
  dplyr::mutate(
    any_complication = TRUE,
    grade3_complication = grade >= 3
  ) \%>\%
  tbl_ae_binary(
    include = c(any_complication, grade3_complication),
    id = patient_id,
    ae = adverse_event,
    soc = system_organ_class,
    include_label =
      list(any_complication = "Any Grade Complication",
           grade3_complication = "Grade 3+ Complication")
  )
}
