% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ae_count.R
\name{tbl_ae_count}
\alias{tbl_ae_count}
\title{Tabulate Raw AE Counts}
\usage{
tbl_ae_count(
  data,
  ae,
  soc = NULL,
  by = NULL,
  strata = NULL,
  by_values = NULL,
  missing_text = "Unknown",
  header = "**{level}**"
)
}
\arguments{
\item{data}{Data frame}

\item{ae}{Variable name of the adverse event column}

\item{soc}{Variable name of the system organ class column}

\item{by}{Variable to split results by, e.g. report AEs by grade}

\item{strata}{Variable to stratify results by, e.g. report AEs summaries
by treatment group}

\item{by_values}{Optional vector of complete by values, listed in desired order,
to achieve correct table structure for the situation in which an adverse
event of a certain grade is not observed for a given soc}

\item{missing_text}{String that will be shown for missing levels of \verb{by=},
Default is \code{"Unknown"}}

\item{header}{String indicating the header to be placed in the table.
Default is \code{"**{level}**"}}
}
\value{
a 'tbl_ae_count' object
}
\description{
Create a table with the number of AE that were reported.
}
\examples{
\dontshow{if (isTRUE(Sys.getenv("NOT_CRAN") \%in\% c("true", ""))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example 1 -----------------------------------------------------------------
tbl_ae_count_ex1 <-
  tbl_ae_count(
    data = df_adverse_events,
    ae = adverse_event,
    soc = system_organ_class,
    strata = trt,
    by = grade,
    header = "**Grade {level}**"
  )
\dontshow{\}) # examplesIf}
}
