% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep4.R
\name{TIC.func}
\alias{TIC.func}
\title{TIC.function}
\usage{
TIC.func(res, TZ, tZ, Ginv = T, Adex = NULL, method = "TIC")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
This function does some internal work
}
