% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzb1.R
\name{tZB1}
\alias{tZB1}
\title{tZB Function (v1)}
\usage{
tZB1(Xs, sYgrid)
}
\arguments{
\item{sYgrid}{}
}
\value{
A giant constraint matrix.
}
\description{
This function makes a giant constraint matrix, with Ygrid the same for every observation. Ygrid has already been processed into sYgrid, i.e. s(Ygrid), which is a matrix nYgrid by nSY.
}
