% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzb_form.R
\name{tzb_form}
\alias{tzb_form}
\title{tZB Function (v2)}
\usage{
tzb_form(Xs, SYfunc, sYfunc, Ygrid, nsim = 29)
}
\arguments{
\item{nsim}{}
}
\description{
This function makes a giant constraint matrix, with Ygrid the same for every observation. Ygrid has already been processed into sYgrid, i.e. s(Ygrid), which is a matrix nYgrid by nSY.
}
