% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_contract.R
\name{v_contract}
\alias{v_contract}
\title{V Contract Function}
\usage{
v_contract(g, buse)
}
\arguments{
\item{buse}{}
}
\value{
Gradient pieces.
}
\description{
This gives the gradient pieces for b for the optimisation algorithm
}
