% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcontract.R
\name{vcontract}
\alias{vcontract}
\title{V Contract Function}
\usage{
vcontract(g, buse)
}
\arguments{
\item{buse}{}
}
\value{
Gradient pieces.
}
\description{
This gives the gradient pieces for b for the optimisation algorithm
}
