% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnss2.R
\name{xnss2}
\alias{xnss2}
\title{XNSS2 Function}
\usage{
xnss2(
  y,
  x,
  info,
  ydf = 4,
  iyknots = NULL,
  addxint = T,
  bstart = NULL,
  lam = 0,
  lam.vec = NULL,
  gam = 0,
  delta = 0,
  pentype = 0,
  yorth = FALSE,
  xorth = FALSE,
  Ysing = FALSE,
  yorder = 4,
  maxit = 300,
  checkg = FALSE,
  checkH = FALSE,
  tval = 0,
  algor = "optim",
  itprint = 1,
  mask = NULL,
  secalc = F,
  HJreturn = F,
  bthresh = 0.001,
  silent = F,
  fastreturn = F,
  reltol = 1e-08,
  diary = NULL,
  cval = 0.01,
  theta = 0.9,
  maxing = FALSE,
  ygrid = seq(min(y), max(y), len = 101),
  btarg = NULL,
  egam = 0,
  etarg = NULL,
  etgam = 0,
  R = NULL
)
}
\arguments{
\item{R}{}
}
\value{

}
\description{
This is a workhorse solver for the problem.
}
