% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpert8d_ss.R
\name{xpert8d_ss}
\alias{xpert8d_ss}
\title{Xpert8d_ss}
\usage{
xpert8d_ss(
  TYX,
  tYX,
  Kscore = 0,
  gam = 0,
  egam = 0,
  lam = 0,
  lam.vec = NULL,
  maxit = 200,
  algor = "ECOS",
  reltol = 1e-04,
  feastol = 1e-04,
  abstol = 1e-04,
  quiet = F,
  zeros = NULL,
  doprimal = FALSE,
  btarg = 0,
  silent = F,
  nXs = NULL,
  nYS = NULL,
  weights = 1,
  cval = 0.1,
  pen = NULL,
  beta2 = F,
  Xs = NULL,
  sYgrid = NULL,
  bounded = F,
  Cbound = 1e+06,
  threshold = 1e-05
)
}
\arguments{
\item{maxit}{This is the maximum number of iterations that will be run.}

\item{algor}{This is the algorithm that will be used to run the optimisation.}

\item{threshold}{}
}
\value{
This is the output of the funciton.
}
\description{
This function does the main optimisation for the GTR.
}
