% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_knots.R
\name{y_knots}
\alias{y_knots}
\title{Y Knots Function}
\usage{
y_knots(y, e0mode = F, ydf, yorder, delta = 1)
}
\arguments{
\item{delta}{}
}
\value{

}
\description{
This function generates the knots for the dependent variable.
}
