% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2_check.R
\name{beta2.check}
\alias{beta2.check}
\title{Beta2 Check}
\usage{
beta2.check(bmat, Xs, nXs, nYS)
}
\arguments{
\item{nYS}{}
}
\value{
The slope and its minimum valye
}
\description{
This function computes the slope and its minimum value.
}
