% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep4.R
\name{beta2.check}
\alias{beta2.check}
\title{beta2.check}
\usage{
beta2.check(bmat, Xs, nXs, nYS)
}
\arguments{
\item{nYS}{}
}
\value{

}
\description{
This function checks what the smallest beta is as well as returning them.
}
