% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/califunc.R
\name{califunc}
\alias{califunc}
\title{Calibration Function}
\usage{
califunc(Y, X, qY = NULL, dgp = "dr", link = "probit")
}
\arguments{
\item{link}{}
}
\value{

}
\description{
This function calibrates coefficients.
}
