% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_contract.R
\name{cm_contract}
\alias{cm_contract}
\title{Contribution Matrix Contraction Function}
\usage{
cm_contract(cm, buse)
}
\arguments{
\item{cm}{The contribution matrix}

\item{buse}{A matrix with the Betas to use.}
}
\description{
This contracts a 'contributions' (n by big) matrix by the mask in buse, by removing the appropriate columns.
}
