% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids_func.R
\name{grids.func}
\alias{grids.func}
\title{Grids Function}
\usage{
grids.func(x, X.type, nxgrid = 101, grid.user = NULL, gridx.cont = F)
}
\arguments{
\item{gridx.cont}{}
}
\value{

}
\description{
This generates the grids of values for the explanatory variable.
}
