% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtreg-package.R
\docType{package}
\name{gtreg-package}
\alias{gtreg}
\alias{gtreg-package}
\title{gtreg: Regulatory Tables For Clinical Research with 'gtsummary'}
\description{
Creates tables suitable for regulatory agency submission using the 'gtsummary' package as the backend. Tables can be exported to HTML, RTF, Word, PDF and more. Highly customizable outputs are available by utilizing existing styling function from 'gtsummary' as well as custom options designed for regulatory tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.pipinghotdata.com/gtreg}
  \item Report bugs at \url{https://github.com/shannonpileggi/gtreg/issues}
}

}
\author{
\strong{Maintainer}: Shannon Pileggi \email{shannon@pipinghotdata.com} (\href{https://orcid.org/0000-0002-7732-4164}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})
}

}
\keyword{internal}
