% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtreg-package.R
\docType{package}
\name{gtreg-package}
\alias{gtreg}
\alias{gtreg-package}
\title{gtreg: 'gt' Backed Regulatory Tables For Clinical Research}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shannonpileggi.github.io/gtreg}
}

}
\author{
\strong{Maintainer}: Shannon Pileggi \email{shannon@pipinghotdata.com} (\href{https://orcid.org/0000-0002-7732-4164}{ORCID})

Authors:
\itemize{
  \item Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})
}

}
\keyword{internal}
