% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtreg-package.R
\docType{package}
\name{gtreg-package}
\alias{gtreg}
\alias{gtreg-package}
\title{gtreg: Regulatory Tables for Clinical Research}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates tables suitable for regulatory agency submission by leveraging the 'gtsummary' package as the back end. Tables can be exported to HTML, Word, PDF and more. Highly customized outputs are available by utilizing existing styling functions from 'gtsummary' as well as custom options designed for regulatory tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shannonpileggi/gtreg}
  \item \url{https://shannonpileggi.github.io/gtreg/}
  \item Report bugs at \url{https://github.com/shannonpileggi/gtreg/issues}
}

}
\author{
\strong{Maintainer}: Shannon Pileggi \email{shannon.pileggi@gmail.com} (\href{https://orcid.org/0000-0002-7732-4164}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})
}

}
\keyword{internal}
