% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep4.R
\name{iyknots.func}
\alias{iyknots.func}
\title{iyknots.func}
\usage{
iyknots.func(y, e0mode = F, ydf, yorder, delta = 1)
}
\arguments{
\item{delta}{}
}
\value{

}
\description{
This function is internal. It generates the knots for the splines.
}
