% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QGM_check.R
\name{QGM.check}
\alias{QGM.check}
\title{Quasi-Gaussian Monotonicity Check}
\usage{
QGM.check(
  y,
  x,
  res,
  xgrid.qgm = seq(min(x), max(x), len = 101),
  ugrid,
  nxgrid = 101,
  nygrid,
  ng.qgm = 201,
  nXs,
  nYS,
  ydf,
  yorder,
  iyknots = NULL,
  info = info,
  Ysing = FALSE,
  delta = 1,
  easy = F,
  zeros = NULL,
  e0mode = T
)
}
\arguments{
\item{e0mode}{}
}
\value{
A design matrix with the result of the check.
}
\description{
This function checks whether there is quasi-Gaussian monotonicity in the gtr result.
}
