% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{vars}
\alias{select}
\alias{where}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{num_range}
\alias{all_of}
\alias{any_of}
\alias{everything}
\alias{last_col}
\alias{one_of}
\alias{as_tibble}
\alias{as_gt}
\alias{as_flex_table}
\alias{as_hux_table}
\alias{as_hux_xlsx}
\alias{as_kable}
\alias{as_kable_extra}
\alias{inline_text}
\alias{add_overall}
\alias{modify_spanning_header}
\alias{modify_header}
\alias{modify_footnote}
\alias{modify_caption}
\alias{modify_column_indent}
\alias{bold_labels}
\alias{bold_levels}
\alias{italicize_labels}
\alias{show_header_names}
\alias{style_number}
\alias{style_sigfig}
\alias{style_percent}
\alias{all_continuous}
\alias{all_continuous2}
\alias{all_categorical}
\alias{all_stat_cols}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}, \code{\link[dplyr:reexports]{all_of}}, \code{\link[dplyr:reexports]{any_of}}, \code{\link[dplyr:reexports]{contains}}, \code{\link[dplyr:reexports]{ends_with}}, \code{\link[dplyr:reexports]{everything}}, \code{\link[dplyr:reexports]{last_col}}, \code{\link[dplyr:reexports]{matches}}, \code{\link[dplyr:reexports]{num_range}}, \code{\link[dplyr:reexports]{one_of}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:reexports]{starts_with}}, \code{\link[dplyr]{vars}}, \code{\link[dplyr:reexports]{where}}}

  \item{gtsummary}{\code{\link[gtsummary]{add_overall}}, \code{\link[gtsummary:select_helpers]{all_categorical}}, \code{\link[gtsummary:select_helpers]{all_continuous}}, \code{\link[gtsummary:select_helpers]{all_continuous2}}, \code{\link[gtsummary:select_helpers]{all_stat_cols}}, \code{\link[gtsummary]{as_flex_table}}, \code{\link[gtsummary]{as_gt}}, \code{\link[gtsummary]{as_hux_table}}, \code{\link[gtsummary:as_hux_table]{as_hux_xlsx}}, \code{\link[gtsummary]{as_kable}}, \code{\link[gtsummary]{as_kable_extra}}, \code{\link[gtsummary:bold_italicize_labels_levels]{bold_labels}}, \code{\link[gtsummary:bold_italicize_labels_levels]{bold_labels}}, \code{\link[gtsummary:bold_italicize_labels_levels]{bold_levels}}, \code{\link[gtsummary]{inline_text}}, \code{\link[gtsummary:bold_italicize_labels_levels]{italicize_labels}}, \code{\link[gtsummary:bold_italicize_labels_levels]{italicize_labels}}, \code{\link[gtsummary]{modify_caption}}, \code{\link[gtsummary]{modify_column_indent}}, \code{\link[gtsummary:deprecated_modify_footnote]{modify_footnote}}, \code{\link[gtsummary:modify]{modify_header}}, \code{\link[gtsummary:modify]{modify_spanning_header}}, \code{\link[gtsummary:modify]{show_header_names}}, \code{\link[gtsummary]{style_number}}, \code{\link[gtsummary]{style_percent}}, \code{\link[gtsummary]{style_sigfig}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

