% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\title{Simulate Data}
\usage{
simdata(
  Ydata,
  Xdata,
  bvec,
  n = 100,
  sde = 1,
  dgp = "dr",
  ugrid = NULL,
  xgrid = NULL
)
}
\arguments{
\item{xgrid}{}
}
\value{
Some fake data.
}
\description{
This function generates some fake data.
}
