% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ae_count.R
\name{tbl_ae_count}
\alias{tbl_ae_count}
\title{Tabulate Raw AE Counts}
\usage{
tbl_ae_count(
  data,
  ae,
  soc = NULL,
  by = NULL,
  strata = NULL,
  by_values = NULL,
  missing_text = "Unknown",
  statistic = "{n}",
  header = "**{level}**"
)
}
\arguments{
\item{data}{Data frame}

\item{ae}{Variable name of the adverse event column}

\item{soc}{Variable name of the system organ class column}

\item{by}{Variable to split results by, e.g. report AEs by grade}

\item{strata}{Variable to stratify results by, e.g. report AEs summaries
by treatment group}

\item{by_values}{Optional vector of complete by values, listed in desired order,
to achieve correct table structure for the situation in which an adverse
event of a certain grade is not observed for a given soc}

\item{missing_text}{String that will be shown for missing levels of \verb{by=},
Default is \code{"Unknown"}}

\item{statistic}{String indicating the statistics that will be reported.
The default is \code{"{n} ({p})"}}

\item{header}{String indicating the header to be placed in the table.
Default is \code{"**{level}**"}}
}
\value{

}
\description{
Tabulate Raw AE Counts
}
\examples{
# add example
}
