% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ae_count.R
\name{tbl_ae_count}
\alias{tbl_ae_count}
\title{Tabulate Raw AE Counts}
\usage{
tbl_ae_count(
  data,
  ae,
  soc = NULL,
  by = NULL,
  strata = NULL,
  by_values = NULL,
  missing_text = "Unknown",
  header_by = NULL,
  zero_symbol = "—",
  digits = NULL,
  sort = c("alphanumeric", "frequency")
)
}
\arguments{
\item{data}{Data frame}

\item{ae}{Variable name of the adverse event column}

\item{soc}{Variable name of the system organ class column}

\item{by}{Variable to split results by, e.g. report AEs by grade}

\item{strata}{Variable to stratify results by, e.g. report AEs summaries
by treatment group}

\item{by_values}{Optional vector of complete by values, listed in desired order,
to achieve correct table structure for the situation in which an adverse
event of a certain grade is not observed for a given soc}

\item{missing_text}{String that will be shown for missing levels of \verb{by=},
Default is \code{"Unknown"}}

\item{header_by}{String indicating the header to be placed in the table.
Default is \code{"**{level}**"} and only used when \verb{by=} is specified.}

\item{zero_symbol}{String used to represent cells with zero counts. Default
is the em-dash (\code{"\\U2014"}). Using \code{zero_symbol = NULL} will print the
zero count statistics, e.g. \code{"0 (0)"}}

\item{digits}{Specifies the number of decimal places to round the summary statistics.
By default integers are shown to zero decimal places, and percentages are
formatted with \code{style_percent()}. If you would like to modify either
of these, pass a vector of integers indicating the number of decimal
places to round the statistics. For example, if the statistic being
calculated is \code{"{n} ({p}\%)"} and you want the percent rounded to
2 decimal places use \code{digits = c(0, 2)}.
User may also pass a styling function: \code{digits = style_sigfig}
@param sort Order AEs and SOCs appear in the tables. Default is
\code{"alphanumeric"}. \code{sort = "frequency"} will sort in decreasing frequency
order.}
}
\value{
a 'tbl_ae_count' object
}
\description{
Create a table with the number of AE that were reported.
}
\examples{
\dontshow{if (isTRUE(Sys.getenv("NOT_CRAN") \%in\% c("true", ""))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example 1 -----------------------------------------------------------------
tbl_ae_count_ex1 <-
  tbl_ae_count(
    data = df_adverse_events,
    ae = adverse_event,
    soc = system_organ_class,
    strata = trt,
    by = grade,
    header_by = "**Grade {level}**"
  )
\dontshow{\}) # examplesIf}
}
